🚚 Move gesture script & files to separate folder
