+++ Assets/Scripts/Gestures/HandShape.cs
@@ -1,0 +1,11 @@
+namespace RogueApeStudios.SecretsOfIgnacios.Gestures
+{
+    public enum HandShape
+    {
+        Start,
+        Fire,
+        Water,
+        Earth,
+        Air
+    }
+}
