+++ Assets/Scripts/Gestures/ScriptableObjects.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ee89e5bf1927639418b664b0d5d0f554
+guid: 0b36811fe882b7f48bdf811d22e165de
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
