+++ Assets/Scripts/Gestures/ScriptableObjects/Testy Boy 2.asset.meta
