--- Assets/Scripts/Spells/HandShape.cs
@@ -1,11 +1,0 @@
-namespace RogueApeStudios.SecretsOfIgnacios.Spells
-{
-    public enum HandShape
-    {
-        Start,
-        Fire, 
-        Water, 
-        Earth, 
-        Air
-    }
-}
