+++ Assets/Scripts/Spells/ScriptableObjects.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0b36811fe882b7f48bdf811d22e165de
+guid: 1f5de21fbdb250249a99893898c3365a
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
