--- Assets/Scripts/Spells/ScriptableObjects/Gestures.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: aea56e967f249e249a112b764ed4ee74
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
