--- Assets/Scripts/Spells/SequenceManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bc079f58c05b87c45829978c2e1a97a3
